cleaned up the controller
